Move hooks into their corresponding feature folder
