+++ src/features/identity/hooks.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { type ManagedIdentity } from "@web5/agent";
 import { IdentityAgentManager } from "@/features/identity/IdentityAgentManager";
-import { type CustomizableUseQueryOptions } from "@/hooks/types";
+import { type CustomizableUseQueryOptions } from "@/types/use-query";
 
 export const useIdentityList = (
   options: CustomizableUseQueryOptions<ManagedIdentity[]> = {}
